Move examples, merge swagger and jdef
