+++ proto/private/j5/schema/v1/schema.proto
@@ -78,7 +78,6 @@ message Enum {
 
 message String {
   optional string format = 1;
-  optional string example = 2;
   Rules rules = 3;
 
   message Rules {
