+++ schema/export/json.go
@@ -1,4 +1,4 @@
-package swagger
+package export
 
 import (
 	"fmt"
